LeetCode/src/com/leetcode/problem219/containsduplicateii/Solution.java
package com.leetcode.problem219.containsduplicateii;

import java.util.HashMap;
import java.util.Map;

/**
 * Given an array of integers and an integer k, find out whether there are two
 * distinct indices i and j in the array such that nums[i] = nums[j] and the
 * difference between i and j is at most k.
 * 
 * @author dongyuxi
 *
 */
public class Solution {
    public boolean containsNearbyDuplicate(int[] nums, int k) {
        if (null == nums || 0 == nums.length) {
            return false;
        }
        Map<Integer, Integer> map = new HashMap<Integer, Integer>();
        for (int i = 0; i < nums.length; i++) {
            if (!map.containsKey(nums[i])) {
                map.put(nums[i], i);
            } else {
                if (i - map.get(nums[i]) <= k) {
                    return true;
                }
                map.put(nums[i], i);
            }
        }
        return false;
    }
}
